+++ tests/testthat/test-05-create_dir.R
@@ -18,8 +18,4 @@ test_that("create_dir fully creates the folder structure", {
 
 })
 
-test_that("verify throws needed errors", {
-  expect_error(verify(main_path("abcdefghijklmop")))
-  expect_error(verify(c(main_path("abcdef"), main_path("ghijklmop"))))
-})